requirements.txt
asyncio==3.4.3
base58==2.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
colored==2.2.4
coloredlog==0.2.5
coloredlogs==15.0.1
ecdsa==0.18.0
hdwallet==2.2.1
humanfriendly==10.0
idna==3.6
mnemonic==0.21
pycryptodome==3.20.0
pyreadline3==3.4.1
requests==2.31.0
six==1.16.0
termcolor==2.4.0
urllib3==2.2.1

pip~=24.0
setuptools~=69.2.0
python-dotenv~=1.0.1